# docs/swagger.yaml
basePath: /api
definitions:
  account.AccountType:
    enum:
    - BANK_ACCOUNT
    - SAVINGS_ACCOUNT
    - CHECKING_ACCOUNT
    - CREDIT_CARD
    - INVESTMENT_ACCOUNT
    - CRYPTO_WALLET
    - CRYPTO_EXCHANGE
    - BROKER
    - PENSION
    - INSURANCE
    - HOME
    - SAFE
    - OTHER
    type: string
    x-enum-varnames:
    - BankAccount
    - SavingsAccount
    - CheckingAccount
    - CreditCard
    - InvestmentAccount
    - CryptoWallet
    - CryptoExchange
    - Broker
    - Pension
    - Insurance
    - Home
    - Safe
    - Other
...
  routes.AccountResponse:
    properties:
      accountType:
        type: string
      balance:
        type: number
      createdAt:
        type: string
      currencyCode:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  routes.AccountSummaryResponse:
    properties:
      byCurrency:
        additionalProperties:
          type: number
        type: object
      byType:
        additionalProperties:
          type: integer
        type: object
      totalAccounts:
        type: integer
      totalBalance:
        type: number
    type: object
  routes.AccountsListResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/routes.AccountResponse'
        type: array
      total:
        type: integer
    type: object
...
  routes.CreateAccountRequest:
    properties:
      accountType:
        $ref: '#/definitions/account.AccountType'
      balance:
        type: number
      currencyCode:
        type: string
      name:
        type: string
    required:
    - accountType
    - currencyCode
    - name
    type: object
...
  routes.UpdateAccountRequest:
    properties:
      accountType:
        $ref: '#/definitions/account.AccountType'
      balance:
        type: number
      currencyCode:
        type: string
      name:
        type: string
    required:
    - accountType
    - currencyCode
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a wallet service API for managing user accounts, assets, and
    transactions
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Wallet Service API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Get all accounts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.AccountsListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all user accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account for the authenticated user
      parameters:
      - description: Account creation data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/routes.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AccountResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing account for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Get a specific account by ID for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AccountResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an existing account for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AccountResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an account
      tags:
      - accounts
  /accounts/summary:
    get:
      consumes:
      - application/json
      description: Get summary statistics for all user accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AccountSummaryResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account summary
      tags:
      - accounts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"